refactor: split lexer into multiple files
