+++ src/error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, thiserror::Error)]
+pub enum Error {}
+
+pub type Result<T> = std::result::Result<T, Error>;
