+++ src/lib.rs
@@ -1,6 +1,7 @@
 mod ast;
 mod builtins;
 mod environment;
+mod error;
 mod evaluator;
 mod lexer;
 mod parser;
@@ -9,6 +10,7 @@ mod repl;
 pub use ast::*;
 pub use builtins::*;
 pub use environment::*;
+pub use error::*;
 pub use evaluator::*;
 pub use lexer::*;
 pub use parser::*;
